codegen: Refactor Event creation
This should make it easier to add Operations which place a fence/barrier
after the last event of the Operation.

In such a case, the call to last_event() should be used to relate
last_event and next_event in the appropriate relation (need to do more
bookkeeping for this, however).