codegen: Backend may be mutable, rename AsmState
Initially, the plan was to maintain any state for more complex ISAs in
AsmState (e.g. registers) -- however, it is cleaner if AsmState just
does what it does now, i.e. allocate events. The job of maintaining
state about the particular backend should be done by the Backend.

This change reflects this intent.